norm fixed/check calloc

+++ Libft/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: mklimina <mklimina@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/26 21:22:38 by maria             #+#    #+#             */
-/*   Updated: 2022/11/11 21:35:54 by mklimina         ###   ########.fr       */
+/*   Updated: 2022/11/22 22:53:30 by mklimina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char * dst, const char *src, size_t dstsize)
+size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
 	unsigned int				dest_len;
 	unsigned int				src_len;
